+++ Apps/Samples/Asteroids/AsteroidsApp.h
@@ -90,7 +90,7 @@ private:
     const float                       m_scene_scale;
     const Constants                   m_scene_constants;
     AsteroidsArray::Settings          m_asteroids_array_settings;
-    size_t                            m_asteroids_complexity = 0;
+    uint32_t                          m_asteroids_complexity          = 0u;
     bool                              m_is_parallel_rendering_enabled = true;
     SceneUniforms                     m_scene_uniforms = { };
